% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2w2015.R
\docType{data}
\name{w2w2015}
\alias{w2w2015}
\title{Wharf to Wharf race data - 2015}
\format{A data frame with fifteen variables: 
  \code{firstname}, \code{lastname}, \code{bib}, \code{city}, \code{state},
  \code{country}, \code{age}, \code{sex}, \code{overall}, \code{oversex},
  \code{overdiv}, \code{elapsed}, \code{elapsedTime}, \code{start},
  \code{startTime}}
\source{
https://github.com/aaronferrucci/wharf2wharf
}
\usage{
data(w2w2015)
}
\description{
Data from the 2015 Wharf to Wharf race in Santa Cruz. This dataset is
filtered and cleaned. The code for scraping, filtering and cleaning 
is here:  https://github.com/aaronferrucci/wharf2wharf
}
\examples{
data(w2w2015)
}
\references{

}
\keyword{datasets}

